# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API POS Seguro",
        "description": "Documentación interactiva de la API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://fictional-couscous-697rjg97wq4vcr579-80.app.github.dev/api",
            "description": "Servidor API Principal"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Iniciar sesión",
                "description": "Autentica al usuario y devuelve un token de Sanctum",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@pos.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Login exitoso"
                                        },
                                        "access_token": {
                                            "type": "string",
                                            "example": "1|abc123def456ghi789..."
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Credenciales incorrectas"
                    }
                }
            }
        },
        "/logout": {
            "post": {
                "tags": [
                    "Autenticación"
                ],
                "summary": "Cerrar sesión",
                "operationId": "68a6108af04af5e3bc1ea68a6c6f1299",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener inventario",
                "description": "Devuelve la lista completa de productos disponibles",
                "operationId": "c1ff6f862214e1896c59cfe0491ce0e8",
                "responses": {
                    "200": {
                        "description": "Lista de productos obtenida correctamente"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Registrar producto",
                "description": "Añade un nuevo producto al catálogo",
                "operationId": "399770d33fa4672f1216707d0e236763",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "name",
                                    "purchase_price",
                                    "sale_price",
                                    "stock"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "PROD-001"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Teclado Mecánico"
                                    },
                                    "purchase_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 500.5
                                    },
                                    "sale_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 850
                                    },
                                    "stock": {
                                        "type": "integer",
                                        "example": 15
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto registrado en el sistema"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/sales": {
            "post": {
                "tags": [
                    "Ventas"
                ],
                "summary": "Procesar nueva venta",
                "description": "Registra una transacción de venta (carrito) de forma atómica y descuenta el stock en tiempo real.",
                "operationId": "0e806786857f022e42a4097cb3e644ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_method",
                                    "items"
                                ],
                                "properties": {
                                    "payment_method": {
                                        "type": "string",
                                        "example": "Tarjeta de Crédito"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "product_id",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "product_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Transacción completada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transacción completada exitosamente."
                                        },
                                        "ticket": {
                                            "properties": {
                                                "folio": {
                                                    "type": "string",
                                                    "example": "TK-A1B2C3D4-1678901234"
                                                },
                                                "fecha": {
                                                    "type": "string",
                                                    "example": "2026-07-25 14:30:00"
                                                },
                                                "cajero": {
                                                    "type": "string",
                                                    "example": "Admin POS"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 1500.5
                                                },
                                                "metodo_pago": {
                                                    "type": "string",
                                                    "example": "Tarjeta de Crédito"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error lógico (ej. stock insuficiente o falla en base de datos)"
                    },
                    "422": {
                        "description": "Error de validación (ej. el producto no existe o formato inválido)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "Token de Acceso",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Autenticación",
            "description": "Autenticación"
        },
        {
            "name": "Productos",
            "description": "Productos"
        },
        {
            "name": "Ventas",
            "description": "Ventas"
        }
    ]
}